refactor: consolida arquivos de tenant context, escopo de modelos e provider tenant-prisma
Move o cls.set('clinicaId', ...) do TenantContextInterceptor pra dentro de
jwt.strategy.ts (usando user.id_Clinica, ja verificado contra o banco, em
vez do clinicaId cru do payload), elimina o interceptor global. Inline
TENANT_SCOPED_MODELS em tenant.extension.ts e TENANT_PRISMA/tenantPrismaProvider
em prisma.module.ts. Puramente estrutural, nenhuma logica muda.
