+++ apps/backend/src/auth/jwt.strategy.ts
@@ -1,8 +1,10 @@
 import { ExtractJwt, Strategy } from 'passport-jwt';
 import { PassportStrategy } from '@nestjs/passport';
 import { Inject, Injectable, UnauthorizedException } from '@nestjs/common';
+import { ClsService } from 'nestjs-cls';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { TokenDto } from 'src/common/dto/token.dto';
+import type { TenantClsStore } from 'src/prisma/tenant.extension';
 import jwtConfig from './config/jwt.config';
 import { ConfigType } from '@nestjs/config';
 
@@ -10,6 +12,7 @@ import { ConfigType } from '@nestjs/config';
 export class JwtStrategy extends PassportStrategy(Strategy) {
     constructor(
         private readonly prisma: PrismaService,
+        private readonly cls: ClsService<TenantClsStore>,
         @Inject(jwtConfig.KEY)
         private readonly jwtConfiguration: ConfigType<typeof jwtConfig>,
     ) {
@@ -35,6 +38,8 @@ export class JwtStrategy extends PassportStrategy(Strategy) {
         if (user.id_Clinica !== payload.clinicaId) {
             throw new UnauthorizedException();
         }
+
+        this.cls.set('clinicaId', user.id_Clinica);
 
         return {
             ...payload,
