+++ apps/backend/src/prisma/prisma.module.ts
@@ -1,6 +1,21 @@
-import { Module } from '@nestjs/common';
+import { Module, Provider } from '@nestjs/common';
+import { ClsService } from 'nestjs-cls';
 import { PrismaService } from './prisma.service';
-import { tenantPrismaProvider } from './tenant-prisma.provider';
+import { tenantExtension, TenantClsStore } from './tenant.extension';
+
+export const TENANT_PRISMA = Symbol('TENANT_PRISMA');
+
+function tenantPrismaFactory(prisma: PrismaService, cls: ClsService<TenantClsStore>) {
+    return tenantExtension(prisma, cls);
+}
+
+export type TenantPrismaClient = ReturnType<typeof tenantPrismaFactory>;
+
+const tenantPrismaProvider: Provider = {
+    provide: TENANT_PRISMA,
+    useFactory: tenantPrismaFactory,
+    inject: [PrismaService, ClsService],
+};
 
 @Module({
     providers: [PrismaService, tenantPrismaProvider],
