--- apps/backend/src/prisma/tenant-scoped-models.ts
@@ -1,7 +1,0 @@
-export const TENANT_SCOPED_MODELS: readonly string[] = [
-    'Usuario',
-    'ListaEspera',
-    'Atendimento',
-    'Prontuario',
-    'LogAuditoria',
-];
